[Nora] Break the auto-updating feature
This change to OCLP is designed to be used on Nora, with Nora-EFI.

This commit disables the ability for OCLP to automatically install
patcher updates from the Dortania repository. If these builds are
installed on my Sonoma partition, they will prevent me from being able
to install root patches as WhateverGreen is disabled if the OS is not
Ventura.