[Nora] Ensure that WebDrivers are not installed
This copy of OCLP is only for use on Nora, paired with Nora-EFI.

This change disables the NVIDIA Web Drivers patching for this branch, so
that I don't need to use an SSDT or DeviceProperties to disable the GTX1080
inside of Nora. This allows me to use different OS versions with different
GPUs, and nothing will conflict.